공부방 페이지 강제 종료 및 사용자 강제 퇴장

+++ frontend/src/pages/SfuPage.tsx
@@ -258,7 +258,10 @@ export default function SfuPage() {
   }, []);
 
   const deleteRoomEvent = () => {
+    if (!window.confirm('방을 삭제하시겠습니까?')) return;
+    alert('방이 삭제되었습니다.');
     if (user) {
+      socket.emit('deleteRoom', roomId);
       deleteRoom({
         studyRoomId: roomId,
       });
@@ -374,7 +377,10 @@ export default function SfuPage() {
       if (!myVideoRef.current) return;
       myVideoRef.current!.srcObject = myStream.current;
 
-      socket.emit(SFU_EVENTS.JOIN, roomId);
+      socket.emit(SFU_EVENTS.JOIN, {
+        userName: user!.userId,
+        studyRoomId: roomId,
+      });
 
       const offer = await createSender();
       socket.emit(SFU_EVENTS.SENDER_OFFER, { offer });
@@ -438,6 +444,12 @@ export default function SfuPage() {
       });
     });
 
+
+    socket.on('deletedThisRoom', () => {
+      alert('방장이 공부방을 삭제했습니다 :(');
+      leaveRoomEvent();
+    });
+    
     // eslint-disable-next-line consistent-return
     return () => {
       socket.off(SFU_EVENTS.CONNECT);
@@ -448,6 +460,7 @@ export default function SfuPage() {
       socket.off(SFU_EVENTS.SENDER_ICECANDIDATE);
       socket.off(SFU_EVENTS.NEW_PEER);
       socket.off(SFU_EVENTS.SOMEONE_LEFT_ROOM);
+
       socket.disconnect();
     };
   }, [screenShare, roomInfo]);
